+++ packages/llamadart_litert_lm_flutter/darwin/llamadart_litert_lm_flutter/Sources/llamadart_litert_lm_flutter/LlamadartLiteRtLmPlugin.swift
@@ -6,6 +6,7 @@ import CLiteRTLM
 #elseif os(macOS)
 import FlutterMacOS
 import Cocoa
+import LiteRtLm
 #endif
 
 public class LlamadartLiteRtLmPlugin: NSObject, FlutterPlugin {